backend/requirements.txt
Django==5.1.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-cors-headers==3.14.0
polars==0.19.1
fastexcel==0.1.4
